BackEnd/requirements.txt
Django>=5.2.7
psycopg2-binary
djangorestframework
django-cors-headers
requests
